[refactor] 리팩토링 및 패키지 구조 수정
