+++ server/src/main/java/com/moovda_project/moovda/module/comment/service/CommentService.java
@@ -65,9 +65,9 @@ public class CommentService {
 
     public void deleteComment(long commentId,long memberId) {
 
-        Comment comment = findVerifiedComment(commentId);
+        Comment comment = findVerifiedComment(commentId); // 인증된 코멘트인지 확인
 
-        checkValidatedMember(memberId,comment);
+        checkValidatedMember(memberId,comment); // 인증된 멤버인지 확인
 
         deleteWatched(commentId);   // 본 영화 목록에서 삭제
 
@@ -89,11 +89,11 @@ public class CommentService {
     }
 
     private void updateStarAvg(Movie movie) {
-        double totalStar = calculateTotalStar(movie);
+        double totalStar = calculateTotalStar(movie); // 총 별점 합
 
-        double averageStar = totalStar / movie.getComments().size();
+        double averageStar = totalStar / movie.getComments().size(); // 별점 합 / 코멘트 개수
 
-        double roundedStar = Double.parseDouble(String.format("%.1f", averageStar));
+        double roundedStar = Double.parseDouble(String.format("%.1f", averageStar)); // 소수점 둘째자리에서 반올림
 
         movie.setStarAvg(roundedStar);
