+++ server/src/main/java/com/moovda_project/moovda/module/genre/entity/MovieGenre.java
@@ -1,6 +1,6 @@
 package com.moovda_project.moovda.module.genre.entity;
 
-import com.moovda_project.moovda.module.genre.entity.Genre;
+
 import com.moovda_project.moovda.module.movie.entity.Movie;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
