+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/MovieSearchCondition.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.module.movie.dto.search;
+package com.moovda_project.moovda.module.movie.dto;
 
 
 import lombok.*;
