+++ server/src/main/java/com/moovda_project/moovda/module/movie/mapper/MovieMapper.java
@@ -4,8 +4,8 @@ import com.moovda_project.moovda.global.dto.PageDto;
 import com.moovda_project.moovda.module.comment.dto.CommentResponseDto;
 import com.moovda_project.moovda.module.comment.entity.Comment;
 import com.moovda_project.moovda.module.movie.dto.*;
-import com.moovda_project.moovda.module.movie.dto.genre.GenreResponseDto;
-import com.moovda_project.moovda.module.movie.dto.staff.StaffResponseDto;
+import com.moovda_project.moovda.module.genre.dto.GenreResponseDto;
+import com.moovda_project.moovda.module.staff.dto.StaffResponseDto;
 import com.moovda_project.moovda.module.movie.entity.Movie;
 
 import com.moovda_project.moovda.module.genre.entity.MovieGenre;
